pybind problems

+++ include/ioh/logger/observer.hpp
@@ -3,8 +3,10 @@
 #include "ioh/common.hpp"
 
 
-namespace ioh {
-    namespace logger {
+namespace ioh
+{
+    namespace logger
+    {
         /**
          * \brief A observer class for \ref ioh::problem::base problems.
          * It contains five types of triggers, which determine when to log a event:
@@ -14,7 +16,8 @@ namespace ioh {
          *	4. Recording evaluations at pre-defined points
          *	5. With a static number for each exponential bucket.
          */
-        class Observer {
+        class Observer
+        {
             /**
              * \brief variable for recording complete optimization process. i.e. every iteration is stored
              */
@@ -57,7 +60,8 @@ namespace ioh {
              * \param exp the exponent
              * \return (int)log(x)/log(b)
              */
-            [[nodiscard]] int compute_base(const size_t evaluations, const int exp) const {
+            [[nodiscard]] int compute_base(const size_t evaluations, const int exp) const
+            {
                 return static_cast<int>(
                     log(evaluations) / log(exp) + 1e-4 // add small const to help with rounding
                 );
@@ -74,14 +78,15 @@ namespace ioh {
                 const common::OptimizationType optimization_type = common::OptimizationType::Minimization,
                 const int trigger_at_time_points_exp_base = 10,
                 const int trigger_at_range_exp_base = 10
-                )
-                : trigger_always_(trigger_always),
-                  trigger_on_improvement_(trigger_on_improvement),
-                  trigger_at_interval_(trigger_on_interval),
-                  trigger_at_time_points_(std::move(trigger_at_time_points)),
-                  trigger_at_time_points_exp_base_(trigger_at_time_points_exp_base),
-                  triggers_per_time_range_(triggers_per_time_range),
-                  triggers_per_time_range_exp_base_(trigger_at_range_exp_base) {
+                ) :
+                trigger_always_(trigger_always),
+                trigger_on_improvement_(trigger_on_improvement),
+                trigger_at_interval_(trigger_on_interval),
+                trigger_at_time_points_(std::move(trigger_at_time_points)),
+                trigger_at_time_points_exp_base_(trigger_at_time_points_exp_base),
+                triggers_per_time_range_(triggers_per_time_range),
+                triggers_per_time_range_exp_base_(trigger_at_range_exp_base)
+            {
                 reset(optimization_type);
             }
 
@@ -89,17 +94,19 @@ namespace ioh {
              * \brief resets \ref current_best_fitness_
              * \param optimization_type Whether minimization or maximization is used
              */
-            void reset(const common::OptimizationType optimization_type) {
+            void reset(const common::OptimizationType optimization_type)
+            {
                 current_best_fitness_ = optimization_type == common::OptimizationType::Maximization
-                                            ? -std::numeric_limits<double>::infinity()
-                                            : std::numeric_limits<double>::infinity();
+                    ? -std::numeric_limits<double>::infinity()
+                    : std::numeric_limits<double>::infinity();
             }
 
             /**
              * \brief returns true when trigger always is set to true
              * \return \ref trigger_always_
              */
-            [[nodiscard]] bool trigger_always() const {
+            [[nodiscard]] bool trigger_always() const
+            {
                 return trigger_always_;
             }
 
@@ -107,7 +114,8 @@ namespace ioh {
              * \brief return true when a trigger at a static interval is defined
              * \return \ref trigger_at_interval_ != 0;
              */
-            [[nodiscard]] bool trigger_at_interval() const {
+            [[nodiscard]] bool trigger_at_interval() const
+            {
                 return trigger_at_interval_ != 0;
             }
 
@@ -115,7 +123,8 @@ namespace ioh {
              * \brief return true when a trigger on objective function improvement is defined
              * \return \ref trigger_on_improvement_
              */
-            [[nodiscard]] bool trigger_on_improvement() const {
+            [[nodiscard]] bool trigger_on_improvement() const
+            {
                 return trigger_on_improvement_;
             }
 
@@ -124,16 +133,18 @@ namespace ioh {
              * \brief returns true when a trigger on specific time points is defined
              * \return true when \ref trigger_at_time_points_ is not empty
              */
-            [[nodiscard]] bool trigger_at_time_points() const {
+            [[nodiscard]] bool trigger_at_time_points() const
+            {
                 return !trigger_at_time_points_.empty() && !(
-                           trigger_at_time_points_.size() == 1 && trigger_at_time_points_[0] == 0);
+                    trigger_at_time_points_.size() == 1 && trigger_at_time_points_[0] == 0);
             }
 
             /**
              * \brief returns true when a trigger on a specific time range is defined
              * \return \ref triggers_per_time_range_ > 0;
              */
-            [[nodiscard]] bool trigger_at_time_range() const {
+            [[nodiscard]] bool trigger_at_time_range() const
+            {
                 return triggers_per_time_range_ > 0;
             }
 
@@ -143,7 +154,8 @@ namespace ioh {
              * \return true when the current number of evaluations is in the specified interval 
              */
             [[nodiscard]] bool
-            interval_trigger(const size_t evaluations) const {
+            interval_trigger(const size_t evaluations) const
+            {
                 return trigger_at_interval() && (evaluations == 1 || evaluations % trigger_at_interval_ == 0);
             }
 
@@ -154,7 +166,8 @@ namespace ioh {
              * \param evaluations the current number of evaluations
              * \return true when the current evaluation should be logged
              */
-            [[nodiscard]] bool time_range_trigger(const size_t evaluations) const {
+            [[nodiscard]] bool time_range_trigger(const size_t evaluations) const
+            {
                 if (!trigger_at_time_range())
                     return false;
 
@@ -164,7 +177,7 @@ namespace ioh {
                 const auto stop = static_cast<int>(std::pow(triggers_per_time_range_exp_base_, base + 1));
                 const auto interval = static_cast<int>(stop / triggers_per_time_range_);
 
-                return  (evaluations - start) % interval == 0;
+                return (evaluations - start) % interval == 0;
             }
 
             /**
@@ -175,14 +188,17 @@ namespace ioh {
              * \param evaluations the current number of evaluations
              * \return true when the current evaluation should be logged
              */
-            [[nodiscard]] bool time_points_trigger(const size_t evaluations) const {
+            [[nodiscard]] bool time_points_trigger(const size_t evaluations) const
+            {
                 if (!trigger_at_time_points())
                     return false;
 
-                const auto factor = std::pow(trigger_at_time_points_exp_base_,
-                                        compute_base(evaluations, trigger_at_time_points_exp_base_));
+                const auto factor = static_cast<size_t>(std::pow(trigger_at_time_points_exp_base_,
+                                                                   compute_base(
+                                                                       evaluations, trigger_at_time_points_exp_base_)));
 
-                for (const auto &e : trigger_at_time_points_) {
+                for (const auto &e : trigger_at_time_points_)
+                {
                     if (evaluations == (e * factor))
                         return true;
                 }
@@ -197,14 +213,15 @@ namespace ioh {
              */
             bool improvement_trigger(
                 const double fitness,
-                const common::OptimizationType optimization_type = common::OptimizationType::Minimization) {
-                if (trigger_on_improvement_ && compare_objectives(fitness, current_best_fitness_, optimization_type)) {
+                const common::OptimizationType optimization_type = common::OptimizationType::Minimization)
+            {
+                if (trigger_on_improvement_ && compare_objectives(fitness, current_best_fitness_, optimization_type))
+                {
                     current_best_fitness_ = fitness;
                     return true;
                 }
                 return false;
             }
         };
-
     }
 }
